Large scale solve solves...don't know how to check answer
The good news is that despite a 1e16 condition number the solve
doesn't look awful.  The bad news is that I don't know how to check
tha answer.  The reason for this is that the TH matrix is highly
degenerate when the material properties in the two regions are
identical.  I think CS has the old analytic solution?