Trying a displacement BC problem.
What's interesting here is that the combined condition number is still
horrible for the interface case, 1e18.  The condition number for the
simple displacement only BC case with no interface is 3 and works
fine.  Something about the boundary assembly makes this problem very
singular.  This suggests a conceptual error on my part.