Corrected indices for location of applied displacment BCs.
This is a big improvement.  The problem was that I was applying
boundary conditions associated with their element numbering.  This
might seem reasonable but for the multi-domain case we introduce the
matching boundary conditions.  With the current construction these are
placed at the top of the design matrix (TH).  This pushes any use
specified boundary conditions down away from their original indices.
Need a good diagram of this and general way of dealing with it automatically.